+++ examples/window.rs
@@ -17,9 +17,11 @@ fn main() {
         .build(&event_loop)
         .unwrap();
 
-    event_loop.run(move |event, _, control_flow| {
+    event_loop.run(move |event, event_loop, control_flow| {
+        dbg!(event_loop.cursor_position().unwrap());
+
         control_flow.set_wait();
-        println!("{:?}", event);
+        // println!("{:?}", event);
 
         match event {
             Event::WindowEvent {