+++ src/event_loop.rs
@@ -15,6 +15,8 @@ use instant::{Duration, Instant};
 use once_cell::sync::OnceCell;
 use raw_window_handle::{HasRawDisplayHandle, RawDisplayHandle};
 
+use crate::dpi::PhysicalPosition;
+use crate::error::ExternalError;
 use crate::{event::Event, monitor::MonitorHandle, platform_impl};
 
 /// Provides a way to retrieve events from the system and from the windows that were registered to
@@ -358,6 +360,16 @@ impl<T> EventLoopWindowTarget<T> {
     pub fn set_device_event_filter(&self, _filter: DeviceEventFilter) {
         #[cfg(any(x11_platform, wayland_platform, windows))]
         self.p.set_device_event_filter(_filter);
+    }
+
+    /// Returns the current cursor position
+    ///
+    /// ## Platform-specific
+    ///
+    /// - **iOS / Android**: Unsupported.
+    #[inline]
+    pub fn cursor_position(&self) -> Result<PhysicalPosition<f64>, ExternalError> {
+        self.p.cursor_position()
     }
 }